refactor(commands): extract webview side effects from auth commands into services
commands/auth.rs shrinks from 2,528 to ~920 lines of thin wrappers, per the
guideline that all side effects (fs, process, network, webviews) live in
services/:

- services/webview_util.rs: shared WEBVIEW_USER_AGENT, WebView2 cookie
  extraction (CookieManager COM), tracking-prevention disable
- services/session_key_fallback.rs: PowerShell/.NET and hidden-WebView2
  session-key bootstrap fallbacks for login and QR start
- services/webview_login.rs: GamePass OAuth window + completion poll,
  regular (帳密) web-login window + cookie-harvest watcher, shared
  finalization (cookie-jar seeding deduplicated into one helper)
- services/recaptcha_window.rs: reCAPTCHA helper window, token delivery
  (fragment poll + IPC fallback), cancelled-vs-delivered close flag
- SessionDto moves to models/session.rs (models own DTOs); system.rs now
  uses the shared user-agent constant

All Tauri command names and signatures are unchanged; behavior is a
mechanical move.
