+++ src-tauri/src/commands/auth.rs
@@ -2,58 +2,15 @@
 //!
 //! Thin wrappers: validate inputs, delegate to core/service, map errors to [`ErrorDto`].
 
-use base64::Engine;
-use std::collections::HashMap;
-use std::sync::atomic::{AtomicBool, Ordering};
-use std::sync::{Mutex, OnceLock};
-
-use tauri::Manager;
 use tauri::State;
 
 use crate::core::auth::{self, SessionAction};
 use crate::core::error::{AppError, AuthError};
 use crate::models::app_state::{AppState, SessionInfo};
 use crate::models::error::ErrorDto;
-use crate::models::session::Session;
+use crate::models::session::SessionDto;
 use crate::services::beanfun_service::{self, QrCodeData, QrPollResult};
-
-type SessionKeyFallbackSender = tokio::sync::oneshot::Sender<WebviewFetchResult>;
-
-static SESSION_KEY_WEBVIEW_RESULTS: OnceLock<Mutex<HashMap<String, SessionKeyFallbackSender>>> =
-    OnceLock::new();
-
-#[derive(Debug, Clone)]
-struct WebviewFetchResult {
-    url: String,
-    html: String,
-}
-
-// ---------------------------------------------------------------------------
-// DTOs returned to the frontend
-// ---------------------------------------------------------------------------
-
-/// Subset of [`Session`] safe to send to the frontend (no refresh token).
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SessionDto {
-    pub session_id: String,
-    pub token: String,
-    pub region: String,
-    pub account_name: String,
-    pub expires_at: String,
-}
-
-impl SessionDto {
-    fn from_session(s: &Session, session_id: &str) -> Self {
-        Self {
-            session_id: session_id.to_string(),
-            token: s.token.clone(),
-            region: format!("{:?}", s.region),
-            account_name: s.account_name.clone(),
-            expires_at: s.expires_at.to_rfc3339(),
-        }
-    }
-}
+use crate::services::{recaptcha_window, session_key_fallback, webview_login};
 
 // ---------------------------------------------------------------------------
 // Session management commands
@@ -101,7 +58,7 @@ pub async fn login(
         login: recaptcha_login,
     };
 
-    let login_result = login_with_native_fallback(
+    let login_result = session_key_fallback::login_with_native_fallback(
         &app,
         &ss.http_client,
         &account,
@@ -303,9 +260,14 @@ pub async fn qr_login_start(
     let ss = state.require_session(&session_id).await?;
     let region = state.config.read().await.region.clone();
 
-    qr_login_start_with_native_fallback(&app, &ss.http_client, &region, &ss.cookie_jar)
-        .await
-        .map_err(login_err_to_dto)
+    session_key_fallback::qr_login_start_with_native_fallback(
+        &app,
+        &ss.http_client,
+        &region,
+        &ss.cookie_jar,
+    )
+    .await
+    .map_err(login_err_to_dto)
 }
 
 /// Poll the status of an in-progress QR-code login.
@@ -799,613 +761,37 @@ fn login_err_to_dto(err: beanfun_service::LoginError) -> ErrorDto {
     }
 }
 
-async fn login_with_native_fallback(
-    app: &tauri::AppHandle,
-    client: &reqwest::Client,
-    account: &str,
-    password: &str,
-    region: &crate::models::session::Region,
-    cookie_jar: &std::sync::Arc<reqwest::cookie::Jar>,
-    tokens: &beanfun_service::RecaptchaTokens,
-) -> Result<Session, beanfun_service::LoginError> {
-    match beanfun_service::login(client, account, password, region, cookie_jar, tokens).await {
-        Ok(session) => Ok(session),
-        Err(err) if should_try_session_key_fallback(&err) => {
-            tracing::warn!(
-                region = ?region,
-                account = %account,
-                "Primary login path failed before session key bootstrap; trying native fallbacks"
-            );
-            let session_key = fetch_session_key_with_fallback(app, region).await?;
-            beanfun_service::login_with_session_key(
-                client,
-                account,
-                password,
-                region,
-                cookie_jar,
-                &session_key,
-                tokens,
-            )
-            .await
-        }
-        Err(err) => Err(err),
-    }
-}
-
-async fn qr_login_start_with_native_fallback(
-    app: &tauri::AppHandle,
-    client: &reqwest::Client,
-    region: &crate::models::session::Region,
-    cookie_jar: &std::sync::Arc<reqwest::cookie::Jar>,
-) -> Result<QrCodeData, beanfun_service::LoginError> {
-    match beanfun_service::qr_login_start(client, region, cookie_jar).await {
-        Ok(data) => Ok(data),
-        Err(err) if should_try_session_key_fallback(&err) => {
-            tracing::warn!(region = ?region, "QR start failed before session key bootstrap; trying native fallbacks");
-            let session_key = fetch_session_key_with_fallback(app, region).await?;
-            beanfun_service::qr_login_start_with_session_key(client, region, &session_key).await
-        }
-        Err(err) => Err(err),
-    }
-}
-
-fn should_try_session_key_fallback(err: &beanfun_service::LoginError) -> bool {
-    matches!(
-        err,
-        beanfun_service::LoginError::Network(
-            crate::core::error::NetworkError::ConnectionFailed { .. }
-        ) | beanfun_service::LoginError::Network(crate::core::error::NetworkError::Timeout { .. })
-    )
-}
-
-async fn fetch_session_key_with_fallback(
-    app: &tauri::AppHandle,
-    region: &crate::models::session::Region,
-) -> Result<String, beanfun_service::LoginError> {
-    if let Some(session_key) = fetch_session_key_via_powershell(region).await? {
-        tracing::info!(region = ?region, session_key_len = session_key.len(), "Session key obtained via PowerShell/.NET fallback");
-        return Ok(session_key);
-    }
-
-    let webview = fetch_session_key_via_webview2(app, region).await?;
-    let session_key = match region {
-        crate::models::session::Region::HK => {
-            beanfun_service::parse_hk_session_key_html(&webview.html)?
-        }
-        crate::models::session::Region::TW => {
-            beanfun_service::parse_tw_session_key_url(&webview.url)?
-        }
-    };
-    tracing::info!(region = ?region, session_key_len = session_key.len(), "Session key obtained via WebView2 fallback");
-    Ok(session_key)
-}
-
-async fn fetch_session_key_via_powershell(
-    region: &crate::models::session::Region,
-) -> Result<Option<String>, beanfun_service::LoginError> {
-    #[cfg(target_os = "windows")]
-    {
-        let script = match region {
-            crate::models::session::Region::HK => format!(
-                r#"$ProgressPreference='SilentlyContinue'
-[Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12 -bor [Net.SecurityProtocolType]::Tls13
-$req = [System.Net.HttpWebRequest]::Create('{0}')
-$req.Method = 'GET'
-$req.UserAgent = '{1}'
-$req.AutomaticDecompression = [System.Net.DecompressionMethods]::GZip -bor [System.Net.DecompressionMethods]::Deflate
-$req.Headers['Accept-Encoding'] = 'identity'
-$resp = $req.GetResponse()
-$reader = New-Object System.IO.StreamReader($resp.GetResponseStream(), [System.Text.Encoding]::UTF8)
-$body = $reader.ReadToEnd()
-$reader.Close()
-$resp.Close()
-[Console]::OutputEncoding = [System.Text.Encoding]::UTF8
-Write-Output $body"#,
-                "https://bfweb.hk.beanfun.com/beanfun_block/bflogin/default.aspx?service=999999_T0",
-                WEBVIEW_USER_AGENT
-            ),
-            crate::models::session::Region::TW => format!(
-                r#"$ProgressPreference='SilentlyContinue'
-[Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12 -bor [Net.SecurityProtocolType]::Tls13
-$req = [System.Net.HttpWebRequest]::Create('{0}')
-$req.Method = 'GET'
-$req.UserAgent = '{1}'
-$req.AllowAutoRedirect = $true
-$req.AutomaticDecompression = [System.Net.DecompressionMethods]::GZip -bor [System.Net.DecompressionMethods]::Deflate
-$req.Headers['Accept-Encoding'] = 'identity'
-$resp = $req.GetResponse()
-$finalUrl = $resp.ResponseUri.AbsoluteUri
-$resp.Close()
-[Console]::OutputEncoding = [System.Text.Encoding]::UTF8
-Write-Output $finalUrl"#,
-                "https://tw.beanfun.com/beanfun_block/bflogin/default.aspx?service=999999_T0",
-                WEBVIEW_USER_AGENT
-            ),
-        };
-
-        let encoded =
-            base64::engine::general_purpose::STANDARD.encode(encode_powershell_script(&script));
-        let output = tokio::task::spawn_blocking(move || {
-            std::process::Command::new("powershell.exe")
-                .args(["-NoProfile", "-NonInteractive", "-EncodedCommand", &encoded])
-                .output()
-        })
-        .await
-        .map_err(|e| {
-            beanfun_service::LoginError::Network(
-                crate::core::error::NetworkError::ConnectionFailed {
-                    url: format!("powershell session-key fallback task failed ({e})"),
-                },
-            )
-        })?
-        .map_err(|e| {
-            beanfun_service::LoginError::Network(
-                crate::core::error::NetworkError::ConnectionFailed {
-                    url: format!("powershell session-key fallback failed to launch ({e})"),
-                },
-            )
-        })?;
-
-        if !output.status.success() {
-            tracing::warn!(
-                region = ?region,
-                status = ?output.status.code(),
-                stderr = %String::from_utf8_lossy(&output.stderr),
-                "PowerShell/.NET session-key fallback failed"
-            );
-            return Ok(None);
-        }
-
-        let stdout = String::from_utf8_lossy(&output.stdout).trim().to_string();
-        if stdout.is_empty() {
-            tracing::warn!(region = ?region, "PowerShell/.NET session-key fallback returned empty output");
-            return Ok(None);
-        }
-
-        let session_key = match region {
-            crate::models::session::Region::HK => {
-                beanfun_service::parse_hk_session_key_html(&stdout)?
-            }
-            crate::models::session::Region::TW => {
-                beanfun_service::parse_tw_session_key_url(&stdout)?
-            }
-        };
-
-        Ok(Some(session_key))
-    }
-
-    #[cfg(not(target_os = "windows"))]
-    {
-        let _ = region;
-        Ok(None)
-    }
-}
-
-#[cfg(target_os = "windows")]
-fn encode_powershell_script(script: &str) -> Vec<u8> {
-    script
-        .encode_utf16()
-        .flat_map(|u| u.to_le_bytes())
-        .collect()
-}
-
-async fn fetch_session_key_via_webview2(
-    app: &tauri::AppHandle,
-    region: &crate::models::session::Region,
-) -> Result<WebviewFetchResult, beanfun_service::LoginError> {
-    let request_id = format!("session-key-{}", uuid::Uuid::new_v4());
-    let label = format!("session-key-webview-{}", uuid::Uuid::new_v4());
-    let results = SESSION_KEY_WEBVIEW_RESULTS.get_or_init(|| Mutex::new(HashMap::new()));
-    let (tx, rx) = tokio::sync::oneshot::channel();
-    results.lock().unwrap().insert(request_id.clone(), tx);
-
-    let start_url = match region {
-        crate::models::session::Region::HK => {
-            "https://bfweb.hk.beanfun.com/beanfun_block/bflogin/default.aspx?service=999999_T0"
-        }
-        crate::models::session::Region::TW => {
-            "https://tw.beanfun.com/beanfun_block/bflogin/default.aspx?service=999999_T0"
-        }
-    };
-
-    let data_dir = app
-        .path()
-        .app_data_dir()
-        .map_err(|e| {
-            beanfun_service::LoginError::Network(
-                crate::core::error::NetworkError::ConnectionFailed {
-                    url: format!("failed to get app data dir for session-key webview ({e})"),
-                },
-            )
-        })?
-        .join("session-key-fallback");
-    let _ = std::fs::create_dir_all(&data_dir);
-
-    let init_script = format!(
-        r#"
-(() => {{
-  const REQUEST_ID = "{}";
-  const START_URL = "{}";
-  if (window.location.href === 'about:blank') {{
-    setTimeout(() => {{ window.location.href = START_URL; }}, 50);
-    return;
-  }}
-  const send = () => {{
-    const html = document.documentElement ? document.documentElement.outerHTML : document.body?.outerHTML || '';
-    if (window.__TAURI_INTERNALS__) {{
-      window.__TAURI_INTERNALS__.invoke('session_key_webview_done', {{
-        requestId: REQUEST_ID,
-        url: window.location.href,
-        html
-      }}).catch(err => console.error('[SessionKeyFallback] invoke failed', err));
-    }}
-  }};
-  if (document.readyState === 'complete' || document.readyState === 'interactive') {{
-    setTimeout(send, 50);
-  }} else {{
-    window.addEventListener('DOMContentLoaded', () => setTimeout(send, 50), {{ once: true }});
-  }}
-}})();
-"#,
-        request_id, start_url
-    );
-
-    let window = tauri::WebviewWindowBuilder::new(
-        app,
-        &label,
-        tauri::WebviewUrl::External("about:blank".parse().unwrap()),
-    )
-    .title("Session Key Fallback")
-    .visible(false)
-    .focused(false)
-    .resizable(false)
-    .inner_size(320.0, 240.0)
-    .data_directory(data_dir)
-    .user_agent(WEBVIEW_USER_AGENT)
-    .additional_browser_args("--disable-blink-features=AutomationControlled --no-sandbox")
-    .initialization_script(&init_script)
-    .build()
-    .map_err(|e| {
-        beanfun_service::LoginError::Network(crate::core::error::NetworkError::ConnectionFailed {
-            url: format!("failed to create session-key fallback webview ({e})"),
-        })
-    })?;
-
-    let result = tokio::time::timeout(std::time::Duration::from_secs(20), rx)
-        .await
-        .map_err(|_| {
-            beanfun_service::LoginError::Network(crate::core::error::NetworkError::Timeout {
-                url: "session-key webview fallback timed out".to_string(),
-            })
-        })?
-        .map_err(|_| {
-            beanfun_service::LoginError::Network(
-                crate::core::error::NetworkError::ConnectionFailed {
-                    url: "session-key webview fallback channel closed".to_string(),
-                },
-            )
-        })?;
-
-    let _ = window.destroy();
-    SESSION_KEY_WEBVIEW_RESULTS
-        .get_or_init(|| Mutex::new(HashMap::new()))
-        .lock()
-        .unwrap()
-        .remove(&request_id);
-
-    Ok(result)
-}
-
+/// Receives the captured page from the hidden session-key fallback webview
+/// (invoked by its init script) and hands it to the waiting fallback.
 #[tauri::command]
 pub async fn session_key_webview_done(
     request_id: String,
     url: String,
     html: String,
 ) -> Result<(), ErrorDto> {
-    tracing::info!(
-        request_id = %request_id,
-        final_url = %url,
-        html_len = html.len(),
-        "Session-key WebView2 fallback captured page"
-    );
-
-    if let Some(sender) = SESSION_KEY_WEBVIEW_RESULTS
-        .get_or_init(|| Mutex::new(HashMap::new()))
-        .lock()
-        .unwrap()
-        .remove(&request_id)
-    {
-        let _ = sender.send(WebviewFetchResult { url, html });
-    }
-
+    session_key_fallback::deliver_webview_result(&request_id, url, html);
     Ok(())
 }
 
 // ---------------------------------------------------------------------------
-// GamePass login commands (TW region)
+// Webview login commands (GamePass / regular web login / reCAPTCHA helper)
 // ---------------------------------------------------------------------------
 
-/// User-Agent for WebView2 windows and HTTP requests.
-const WEBVIEW_USER_AGENT: &str = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/134.0.0.0 Safari/537.36";
-
-/// Open GamePass login popup for TW region.
-///
-/// Creates a new session for this GamePass login flow. The entire login
-/// flow happens inside the WebView2:
-/// 1. Navigate to `bflogin/default.aspx` → server redirects to `Login/Index?pSKey={skey}`
-/// 2. Init script auto-clicks `a.use-gama-pass` on the login page
-/// 3. User completes GamePass OAuth in the webview
-/// 4. Init script polls `echo_token.ashx` until session is ready
-/// 5. Fetches account list HTML inside the webview, then signals backend
+/// Open the GamePass login popup (TW region). Returns the new session's ID so
+/// the frontend can track this GamePass session.
 #[tauri::command]
 pub async fn open_gamepass_login(
     app: tauri::AppHandle,
     state: State<'_, AppState>,
 ) -> Result<String, ErrorDto> {
-    use tauri::WebviewWindowBuilder;
-
-    let label = "gamepass-login";
-
-    if let Some(existing) = app.get_webview_window(label) {
-        let _ = existing.destroy();
-        tokio::time::sleep(std::time::Duration::from_millis(300)).await;
-    }
-
-    let config = state.config.read().await;
-    if config.region != crate::models::session::Region::TW {
-        return Err(ErrorDto {
-            code: "AUTH_GAMEPASS_UNSUPPORTED".to_string(),
-            message: "GamePass login is only available for TW region".to_string(),
-            category: crate::models::error::ErrorCategory::Authentication,
-            details: None,
-        });
-    }
-    let incognito = config.gamepass_incognito;
-    drop(config);
-
-    // Create a new session for this GamePass login flow
-    let (session_id, _) = state.create_session().await;
-    tracing::info!("GamePass: created session {session_id}");
-
-    // Fresh flow — clear the "already finalized" guard.
-    GAMEPASS_DONE.store(false, Ordering::SeqCst);
-
-    let data_dir = app.path().app_data_dir().map_err(|e| ErrorDto {
-        code: "SYS_PATH_ERROR".to_string(),
-        message: format!("Failed to get app data dir: {e}"),
-        category: crate::models::error::ErrorCategory::Process,
-        details: None,
-    })?;
-
-    // Navigate to bflogin/default.aspx — the server will redirect to
-    // login.beanfun.com/Login/Index?pSKey={skey} automatically.
-    // This way the WebView2 has the same session cookies as the skey request.
-    let start_url = "https://tw.beanfun.com/beanfun_block/bflogin/default.aspx?service=999999_T0";
-
-    // Pass session_id to the init script so it can invoke gamepass_webview_done with it
-    let init_script = format!(
-        r#"
-    (() => {{
-        const SESSION_ID = "{}";
-        const url = window.location.href;
-        const onBeanfun = url.includes('beanfun.com');
-        const onLoginPage = url.includes('Login/Index');
-        const onGamania = url.includes('accounts.gamania.com');
-        const onDefaultAspx = url.includes('bflogin/default.aspx');
-
-        Object.defineProperty(navigator, 'webdriver', {{get: () => false}});
-
-        // Skip: initial redirect page, gamania OAuth, non-beanfun pages
-        if (onDefaultAspx || onGamania || !onBeanfun) return;
-
-        const _origFetch = window.fetch;
-
-        // On login page: auto-click GamePass button
-        if (onLoginPage) {{
-            function tryClick() {{
-                const btn = document.querySelector('a.use-gama-pass');
-                if (btn) {{ btn.click(); console.log('[GamePass] clicked use-gama-pass'); return true; }}
-                return false;
-            }}
-            // Try immediately
-            if (!tryClick()) {{
-                // Retry with MutationObserver + polling fallback
-                const obs = new MutationObserver(() => {{ if (tryClick()) obs.disconnect(); }});
-                if (document.body) {{
-                    obs.observe(document.body, {{ childList: true, subtree: true }});
-                }}
-                // Also poll every 200ms for up to 10s as fallback
-                let attempts = 0;
-                const poller = setInterval(() => {{
-                    attempts++;
-                    if (tryClick() || attempts > 50) {{
-                        clearInterval(poller);
-                        obs.disconnect();
-                    }}
-                }}, 200);
-            }}
-            return;
-        }}
-
-        // On beanfun.com post-login pages (return.aspx, index.aspx, etc)
-        // Poll echo_token, then fetch accounts, then signal backend
-        console.log('[GamePass] post-login page detected:', url);
-
-        (async function() {{
-            // Poll echo_token until session is confirmed
-            let ready = false;
-            for (let i = 0; i < 60; i++) {{
-                try {{
-                    const r = await _origFetch(
-                        'https://tw.beanfun.com/beanfun_block/generic_handlers/echo_token.ashx?webtoken=1',
-                        {{ credentials: 'include' }}
-                    );
-                    const t = await r.text();
-                    if (t.includes('ResultCode:1')) {{
-                        console.log('[GamePass] session ready at attempt', i);
-                        ready = true;
-                        break;
-                    }}
-                }} catch(e) {{}}
-                await new Promise(r => setTimeout(r, 500));
-            }}
-
-            if (!ready) {{
-                console.log('[GamePass] session not ready, skipping (might be pre-login page)');
-                return;
-            }}
-
-            // Session is ready — fetch account list
-            console.log('[GamePass] fetching account list...');
-            let accountHtml = '';
-            try {{
-                const sc = '610074', sr = 'T9';
-                await _origFetch(
-                    'https://tw.beanfun.com/beanfun_block/auth.aspx?channel=game_zone'
-                    + '&page_and_query=game_start.aspx%3Fservice_code_and_region%3D' + sc + '_' + sr
-                    + '&web_token=1',
-                    {{ credentials: 'include' }}
-                );
-                const listResp = await _origFetch(
-                    'https://tw.beanfun.com/beanfun_block/game_zone/game_server_account_list.aspx'
-                    + '?sc=' + sc + '&sr=' + sr + '&dt=' + Date.now(),
-                    {{ credentials: 'include' }}
-                );
-                accountHtml = await listResp.text();
-                console.log('[GamePass] account list length:', accountHtml.length);
-            }} catch(e) {{
-                console.error('[GamePass] account list fetch failed:', e);
-            }}
-
-            const cookies = document.cookie;
-            let webToken = 'cookie_auth';
-            cookies.split(';').forEach(c => {{
-                const t = c.trim();
-                if (t.startsWith('bfWebToken=')) webToken = t.substring(11);
-            }});
-
-            if (window.__TAURI_INTERNALS__) {{
-                console.log('[GamePass] invoking backend...');
-                window.__TAURI_INTERNALS__.invoke('gamepass_webview_done', {{
-                    sessionId: SESSION_ID,
-                    webToken: webToken,
-                    cookies: cookies,
-                    accountHtml: accountHtml
-                }}).then(() => console.log('[GamePass] SUCCESS'))
-                  .catch(e => console.error('[GamePass] FAILED', e));
-            }}
-        }})();
-    }})();
-    "#,
-        session_id
-    );
-
-    let mut builder = WebviewWindowBuilder::new(
-        &app,
-        label,
-        tauri::WebviewUrl::External(start_url.parse().unwrap()),
-    )
-    .title("Beanfun GamePass Login")
-    .inner_size(420.0, 580.0)
-    .min_inner_size(380.0, 500.0)
-    .decorations(true)
-    .resizable(true)
-    .center()
-    .visible(true)
-    .user_agent(WEBVIEW_USER_AGENT)
-    .additional_browser_args("--disable-blink-features=AutomationControlled --no-sandbox")
-    .initialization_script(&init_script)
-    .devtools(true);
-
-    // In normal mode, persist WebView2 data so "remember me" works.
-    // In incognito mode, use a unique temp directory per session.
-    let webview_data_dir = if incognito {
-        let temp = std::env::temp_dir()
-            .join("MapleLink")
-            .join("gamepass-incognito")
-            .join(format!("{}", std::process::id()));
-        let _ = std::fs::create_dir_all(&temp);
-        temp
-    } else {
-        data_dir
-    };
-    builder = builder.data_directory(webview_data_dir.clone());
-
-    let window = builder.build().map_err(|e| ErrorDto {
-        code: "AUTH_GAMEPASS_WINDOW_FAILED".to_string(),
-        message: format!("Failed to open GamePass login window: {e}"),
-        category: crate::models::error::ErrorCategory::Process,
-        details: None,
-    })?;
-
-    // Google / Facebook / Apple sign-in need their third-party storage and an
-    // OAuth popup. WebView2 Tracking Prevention blocks the SDK storage (see the
-    // "Tracking Prevention blocked ... connect.facebook.net / apis.google.com"
-    // errors) and the popup is blocked (POPUP_MAYBE_BLOCKED_OAUTH). Disable
-    // tracking prevention and route window.open through the main window so the
-    // provider login page loads instead of being blocked.
-    disable_tracking_prevention(&window);
-    // Let OAuth popups open as real popup windows (with window.opener) so
-    // Google/Facebook/Apple sign-in can postMessage + close back to the opener.
-    // Navigating the main window instead breaks that final step.
-    if let Err(e) = crate::services::cookie_native::register_native_popup_handler(&window) {
-        tracing::warn!("GamePass: failed to register native popup handler: {e}");
-    }
-
-    // Backend completion poll (the robust path). The injected JS tries to reach
-    // us over IPC, but beanfun's page CSP blocks ipc.localhost, so that invoke
-    // can silently fail and the login stalls on the beanfun portal without ever
-    // reaching the account list. Instead, poll the WebView2 cookie store from the
-    // backend and finalize as soon as the HttpOnly `bfWebToken` appears — exactly
-    // when the OAuth redirect lands back on the beanfun portal.
-    let poll_app = app.clone();
-    let poll_sid = session_id.clone();
-    tauri::async_runtime::spawn(async move {
-        // ~5 minutes at 500ms; the window-gone check ends it early on close.
-        for _ in 0..600 {
-            tokio::time::sleep(std::time::Duration::from_millis(500)).await;
-            if poll_app.get_webview_window("gamepass-login").is_none() {
-                return; // window closed (finalized elsewhere or user cancelled)
-            }
-            let st = poll_app.state::<AppState>();
-            match try_finalize_gamepass(&poll_app, st.inner(), &poll_sid, "", "").await {
-                Ok(true) => return, // finalized
-                Ok(false) => {}     // token not there yet — keep polling
-                Err(e) => tracing::debug!("GamePass poll: {}", e.message),
-            }
-        }
-        tracing::warn!("GamePass: completion poll timed out after ~5 min");
-    });
-
-    // Clean up incognito temp dir when window closes (best-effort)
-    if incognito {
-        let app_clone = app.clone();
-        let temp_dir = webview_data_dir;
-        tauri::async_runtime::spawn(async move {
-            // Wait for window to be destroyed
-            loop {
-                tokio::time::sleep(std::time::Duration::from_secs(1)).await;
-                if app_clone.get_webview_window("gamepass-login").is_none() {
-                    // Small delay to let WebView2 release file locks
-                    tokio::time::sleep(std::time::Duration::from_secs(2)).await;
-                    let _ = std::fs::remove_dir_all(&temp_dir);
-                    tracing::debug!("GamePass: cleaned up incognito temp dir");
-                    break;
-                }
-            }
-        });
-    }
-
-    tracing::info!("GamePass login window opened");
-    // Return the session_id so the frontend can track this GamePass session
-    Ok(session_id)
+    webview_login::open_gamepass_login_window(app, state.inner()).await
 }
 
 /// Called by the GamePass webview init script when login completes.
 ///
-/// Uses WebView2 CookieManager to extract ALL
-/// cookies (including HttpOnly), injects them into the session's reqwest cookie jar,
-/// then fetches game accounts via reqwest.
+/// `web_token` is the JS `document.cookie` value, only useful as a fallback
+/// (bfWebToken is HttpOnly so JS usually can't read it); the finalizer prefers
+/// the WebView2 CookieManager-extracted cookie.
 #[tauri::command]
 pub async fn gamepass_webview_done(
     session_id: String,
@@ -1416,269 +802,19 @@ pub async fn gamepass_webview_done(
     state: State<'_, AppState>,
 ) -> Result<(), ErrorDto> {
     tracing::info!("=== GamePass webview_done (JS IPC) session: {session_id} ===");
-    // Delegate to the shared finalizer. `web_token` is the JS `document.cookie`
-    // value, only useful as a fallback (bfWebToken is HttpOnly so JS usually
-    // can't read it); the finalizer prefers the COM-extracted cookie.
-    let completed =
-        try_finalize_gamepass(&app, state.inner(), &session_id, &account_html, &web_token).await?;
+    let completed = webview_login::try_finalize_gamepass(
+        &app,
+        state.inner(),
+        &session_id,
+        &account_html,
+        &web_token,
+    )
+    .await?;
     if !completed {
         tracing::info!("GamePass (JS IPC): bfWebToken not present yet — backend poll will retry");
     }
     Ok(())
 }
-
-/// Core GamePass completion, shared by the JS `gamepass_webview_done` IPC and the
-/// backend cookie poll. Pulls ALL WebView2 cookies (incl. HttpOnly `bfWebToken`),
-/// seeds the session jar, fetches accounts, emits `gamepass-login-complete`, and
-/// closes the window.
-///
-/// Returns `Ok(true)` once `bfWebToken` exists and the session is installed, or
-/// `Ok(false)` while it isn't there yet (the caller should keep waiting). The
-/// `GAMEPASS_DONE` guard ensures only the first caller to see the token finalizes.
-async fn try_finalize_gamepass(
-    app: &tauri::AppHandle,
-    state: &AppState,
-    session_id: &str,
-    account_html: &str,
-    js_web_token: &str,
-) -> Result<bool, ErrorDto> {
-    use tauri::Emitter;
-
-    let ss = state.require_session(session_id).await?;
-
-    // Extract ALL cookies from WebView2 via CookieManager (including HttpOnly).
-    let all_cookies = extract_webview2_cookies(app, "gamepass-login").await;
-
-    // Find bfWebToken from the extracted cookies (fall back to the JS value only
-    // when it's a real token, not the "cookie_auth" placeholder).
-    let real_web_token = all_cookies
-        .iter()
-        .find(|(name, _, _, _)| name == "bfWebToken")
-        .map(|(_, value, _, _)| value.clone())
-        .filter(|v| !v.is_empty())
-        .or_else(|| {
-            if js_web_token != "cookie_auth" && !js_web_token.is_empty() {
-                Some(js_web_token.to_string())
-            } else {
-                None
-            }
-        });
-
-    let Some(real_web_token) = real_web_token else {
-        // Not logged in yet — no token on any origin. Keep waiting.
-        return Ok(false);
-    };
-
-    // Only the first caller (poll vs JS IPC) that sees the token finalizes.
-    if GAMEPASS_DONE
-        .compare_exchange(false, true, Ordering::SeqCst, Ordering::SeqCst)
-        .is_err()
-    {
-        return Ok(true);
-    }
-
-    tracing::info!(
-        "GamePass: finalizing — {} cookies, bfWebToken = {}...",
-        all_cookies.len(),
-        &real_web_token[..real_web_token.len().min(20)]
-    );
-
-    // Inject ALL cookies into the session's reqwest jar
-    let tw_url: url::Url = "https://tw.beanfun.com/".parse().unwrap();
-    let login_url: url::Url = "https://login.beanfun.com/".parse().unwrap();
-    let newlogin_url: url::Url = "https://tw.newlogin.beanfun.com/".parse().unwrap();
-
-    for (name, value, domain, path) in &all_cookies {
-        let clean_domain = domain.trim_start_matches('.');
-        let jar_url =
-            if clean_domain.contains("login.beanfun.com") && !clean_domain.contains("newlogin") {
-                &login_url
-            } else if clean_domain.contains("newlogin") {
-                &newlogin_url
-            } else {
-                &tw_url
-            };
-        let path_str = if path.is_empty() { "/" } else { path.as_str() };
-        let cookie_str = format!("{}={}; Domain={}; Path={}", name, value, domain, path_str);
-        ss.cookie_jar.add_cookie_str(&cookie_str, jar_url);
-    }
-
-    tracing::info!("GamePass: injected all cookies into session's reqwest jar");
-
-    // Step 3: Build session with real bfWebToken
-    let session = crate::models::session::Session {
-        token: real_web_token,
-        refresh_token: None,
-        expires_at: chrono::Utc::now() + chrono::Duration::hours(6),
-        region: crate::models::session::Region::TW,
-        account_name: "GamePass".to_string(),
-        session_key: None,
-        totp_state: None,
-    };
-
-    // Step 4: Fetch game accounts via reqwest (now has full cookies)
-    let accounts = crate::services::beanfun_service::get_game_accounts(
-        &ss.http_client,
-        &session,
-        &ss.cookie_jar,
-    )
-    .await
-    .unwrap_or_else(|e| {
-        tracing::warn!("GamePass: reqwest get_game_accounts failed: {e}, trying webview HTML");
-        crate::services::beanfun_service::parse_tw_account_list_html(account_html)
-    });
-
-    tracing::info!("GamePass: got {} accounts", accounts.len());
-
-    let dto = SessionDto::from_session(&session, session_id);
-    *ss.session.write().await = Some(session);
-    *ss.game_accounts.write().await = accounts;
-
-    let _ = app.emit("gamepass-login-complete", dto);
-
-    if let Some(win) = app.get_webview_window("gamepass-login") {
-        let _ = win.destroy();
-    }
-
-    tracing::info!("=== GamePass finalized (session installed) ===");
-    Ok(true)
-}
-
-// ---------------------------------------------------------------------------
-// Regular (帳密) web login — full login in a webview, then harvest cookies
-// ---------------------------------------------------------------------------
-//
-// Same robust pattern as GamePass: the user completes the entire login inside a
-// real beanfun page (account + password + reCAPTCHA + any advance check), then
-// we harvest ALL cookies from WebView2 and build the reqwest session — instead
-// of the fragile "grab a reCAPTCHA token and replay it" approach. The only
-// difference from GamePass is the login page prefills the saved credentials.
-
-/// Injected into the regular web-login window. Reads `window.__ML_*` globals
-/// (set by a prelude) — a plain raw string so the JS braces need no escaping.
-const REGULAR_LOGIN_SCRIPT: &str = r#"
-(() => {
-  const SESSION_ID = window.__ML_SESSION_ID__ || "";
-  const ACCOUNT = window.__ML_ACCOUNT__ || "";
-  const PASSWORD = window.__ML_PASSWORD__ || "";
-
-  const url = window.location.href;
-  const onBeanfun = url.includes('beanfun.com');
-  const onLoginPage = url.includes('Login/Index');
-  const onGamania = url.includes('accounts.gamania.com');
-  const onDefaultAspx = url.includes('bflogin/default.aspx');
-
-  try { Object.defineProperty(navigator, 'webdriver', { get: () => false }); } catch (e) {}
-
-  // Tauri hijacks window.alert to its dialog plugin, which isn't permitted for
-  // this remote origin (CSP/ACL) and throws. Route beanfun's alerts to the
-  // console so they never crash the page — real errors also show inline.
-  try { window.alert = function (m) { console.log('[beanfun alert]', m); }; } catch (e) {}
-
-  if (onDefaultAspx || onGamania || !onBeanfun) return;
-
-  const _origFetch = window.fetch;
-  const hasSession = () => document.cookie.indexOf('bfWebToken=') !== -1;
-  console.log('[WebLogin] page', url, '| login?', onLoginPage, '| session?', hasSession());
-
-  // Only treat as the login form when we're on Login/Index AND not yet logged
-  // in — a post-login redirect back to a Login/Index URL should fall through to
-  // the harvest below.
-  if (onLoginPage && !hasSession()) {
-    // Self-heal reCAPTCHA: WebView2 Tracking Prevention is only turned off a
-    // moment AFTER this window is created, so the very first load inits reCAPTCHA
-    // with google/gstatic storage blocked and it fails to render. Wait ~4s (long
-    // enough for prevention to be applied), and if no reCAPTCHA widget appeared,
-    // reload once — the reload runs with prevention off and it renders. Same
-    // proven approach as the standalone reCAPTCHA helper window.
-    try {
-      const store = window.sessionStorage;
-      const RK = '__wl_reloads__';
-      setTimeout(() => {
-        if (document.querySelector('iframe[src*="recaptcha"]')) return; // rendered OK
-        let done = 0;
-        try { done = parseInt(store.getItem(RK) || '0', 10) || 0; } catch (e) {}
-        if (done >= 1) {
-          console.warn('[WebLogin] reCAPTCHA still missing after reload');
-          return;
-        }
-        try { store.setItem(RK, String(done + 1)); } catch (e) {}
-        console.warn('[WebLogin] reCAPTCHA not rendered — reloading once');
-        location.reload();
-      }, 2500);
-    } catch (e) {}
-
-    // beanfun's login is a Vue app with a TWO-STEP flow: enter account →
-    // CheckAccountType (reveals the password field) → enter password →
-    // AccountLogin. The inputs use deliberately-obfuscated names — account is
-    // name="aaa", password is name="inputName" (password only shows after
-    // step 1). We fill each once when it appears (empty), dispatching a native
-    // 'input' event so Vue's v-model picks up the value; the user just solves
-    // the reCAPTCHA and clicks through.
-    const setVal = (el, v) => {
-      el.value = v;
-      el.dispatchEvent(new Event('input', { bubbles: true }));
-      el.dispatchEvent(new Event('change', { bubbles: true }));
-    };
-    let accFilled = false;
-    let pwFilled = false;
-    const prefill = () => {
-      const acc = document.querySelector(
-        'input[name="aaa"], input[type="email"], input[type="text"]:not([type="hidden"])'
-      );
-      const pw = document.querySelector('input[name="inputName"], input[type="password"]');
-      if (acc && ACCOUNT && !accFilled && !acc.value) { setVal(acc, ACCOUNT); accFilled = true; }
-      if (pw && PASSWORD && !pwFilled && !pw.value) { setVal(pw, PASSWORD); pwFilled = true; }
-      return accFilled && pwFilled;
-    };
-    prefill();
-    // Keep watching for up to ~60s — the password field only appears after the
-    // account step (which needs the reCAPTCHA + a click).
-    let n = 0;
-    const timer = setInterval(() => {
-      n++;
-      if (prefill() || n > 120) clearInterval(timer);
-    }, 500);
-
-    // Auto-click the login buttons once the user solves the reCAPTCHA, so they
-    // only ever have to solve the challenge. Step 1 button is "登入帳號"
-    // (submitAccountName), step 2 is "繼續" (submitPassword); both require the
-    // reCAPTCHA token, and it resets between steps (so a fresh token = a fresh
-    // click). We click the first visible, enabled one that matches.
-    const recaptchaToken = () => {
-      try {
-        const g = window.grecaptcha;
-        if (g && g.enterprise && g.enterprise.getResponse) return g.enterprise.getResponse() || '';
-        if (g && g.getResponse) return g.getResponse() || '';
-      } catch (e) {}
-      return '';
-    };
-    let lastToken = '';
-    setInterval(() => {
-      const token = recaptchaToken();
-      if (!token || token === lastToken) return;
-      const btns = document.querySelectorAll('a.ui-btn');
-      for (const b of btns) {
-        const txt = (b.textContent || '').trim();
-        const visible = b.offsetParent !== null;
-        const disabled = b.classList.contains('disabled');
-        if (visible && !disabled && (txt.indexOf('登入帳號') !== -1 || txt.indexOf('繼續') !== -1)) {
-          lastToken = token;
-          console.log('[WebLogin] reCAPTCHA solved — auto-clicking', txt);
-          b.click();
-          break;
-        }
-      }
-    }, 400);
-    return;
-  }
-
-  // Post-login: nothing to do here. beanfun's page CSP blocks Tauri IPC on
-  // these pages, so the backend watches this window's cookies for bfWebToken and
-  // harvests them itself (see open_regular_web_login).
-  console.log('[WebLogin] post-login page — backend will harvest cookies');
-})();
-"#;
 
 /// Open the regular (帳密) web-login window: the user completes the whole login
 /// on the official page (credentials prefilled), then cookies are harvested.
@@ -1690,137 +826,12 @@ pub async fn open_regular_web_login(
     app: tauri::AppHandle,
     state: State<'_, AppState>,
 ) -> Result<(), ErrorDto> {
-    use tauri::WebviewWindowBuilder;
-
     auth::validate_input("account", &account).map_err(to_dto)?;
     auth::validate_input("password", &password).map_err(to_dto)?;
     // Session is pre-created by the frontend.
     let _ = state.require_session(&session_id).await?;
 
-    let label = "web-login";
-    if let Some(existing) = app.get_webview_window(label) {
-        let _ = existing.destroy();
-        tokio::time::sleep(std::time::Duration::from_millis(200)).await;
-    }
-
-    // Fresh incognito profile each time so the login form (for prefill + the
-    // user's reCAPTCHA) always shows instead of auto-logging-in from a cookie.
-    let data_dir = std::env::temp_dir()
-        .join("MapleLink")
-        .join("web-login")
-        .join(uuid::Uuid::new_v4().to_string());
-    let _ = std::fs::create_dir_all(&data_dir);
-    let cleanup_dir = data_dir.clone();
-
-    let start_url = "https://tw.beanfun.com/beanfun_block/bflogin/default.aspx?service=999999_T0";
-
-    let prelude = format!(
-        "window.__ML_SESSION_ID__={};window.__ML_ACCOUNT__={};window.__ML_PASSWORD__={};\n",
-        serde_json::to_string(&session_id).unwrap_or_else(|_| "\"\"".into()),
-        serde_json::to_string(&account).unwrap_or_else(|_| "\"\"".into()),
-        serde_json::to_string(&password).unwrap_or_else(|_| "\"\"".into()),
-    );
-    let init_script = format!("{prelude}{REGULAR_LOGIN_SCRIPT}");
-
-    let window = WebviewWindowBuilder::new(
-        &app,
-        label,
-        tauri::WebviewUrl::External(start_url.parse().expect("static login URL is valid")),
-    )
-    .title("Beanfun 登入")
-    // Wide enough for beanfun's desktop login layout (promo panel + the form
-    // with the account/password fields + reCAPTCHA on the right). At ~420px only
-    // the left promo shows and the form is scrolled off.
-    .inner_size(1000.0, 680.0)
-    .min_inner_size(720.0, 560.0)
-    .center()
-    .visible(true)
-    .user_agent(WEBVIEW_USER_AGENT)
-    .additional_browser_args("--disable-blink-features=AutomationControlled --no-sandbox")
-    .data_directory(data_dir)
-    .initialization_script(&init_script)
-    .devtools(true)
-    .build()
-    .map_err(|e| ErrorDto {
-        code: "AUTH_WEB_LOGIN_WINDOW_FAILED".to_string(),
-        message: format!("Failed to open login window: {e}"),
-        category: crate::models::error::ErrorCategory::Process,
-        details: None,
-    })?;
-
-    // The login page renders Google reCAPTCHA; WebView2 Tracking Prevention
-    // blocks its google.com/gstatic storage and makes it unclickable. Disable it
-    // for this window (same as the standalone reCAPTCHA helper).
-    disable_tracking_prevention(&window);
-
-    // Clean up the incognito profile after the window closes.
-    let cleanup_app = app.clone();
-    tauri::async_runtime::spawn(async move {
-        loop {
-            tokio::time::sleep(std::time::Duration::from_secs(1)).await;
-            if cleanup_app.get_webview_window("web-login").is_none() {
-                tokio::time::sleep(std::time::Duration::from_secs(2)).await;
-                let _ = std::fs::remove_dir_all(&cleanup_dir);
-                break;
-            }
-        }
-    });
-
-    // Watch the window's cookies for bfWebToken (set once login completes) and
-    // harvest them ourselves — beanfun's page CSP blocks Tauri IPC, so the page
-    // can't signal us. Fully URL-agnostic; no IPC/capability needed.
-    let watch_app = app.clone();
-    let watch_session = session_id.clone();
-    let watch_account = account.clone();
-    tauri::async_runtime::spawn(async move {
-        use tauri::Emitter;
-        // ~5 minutes at 1.5s.
-        for _ in 0..200 {
-            tokio::time::sleep(std::time::Duration::from_millis(1500)).await;
-            let Some(_win) = watch_app.get_webview_window("web-login") else {
-                return; // window closed / cancelled
-            };
-            let cookies = extract_webview2_cookies(&watch_app, "web-login").await;
-            if !cookies.iter().any(|(name, _, _, _)| name == "bfWebToken") {
-                continue; // not logged in yet
-            }
-            tracing::info!("web-login: bfWebToken detected — harvesting session");
-            let Some(state) = watch_app.try_state::<AppState>() else {
-                return;
-            };
-            let Some(ss) = state.get_session(&watch_session).await else {
-                return;
-            };
-            match finalize_webview_login(
-                &watch_app,
-                &ss,
-                &watch_session,
-                "web-login",
-                &watch_account,
-                "cookie_auth",
-                "",
-            )
-            .await
-            {
-                Ok(dto) => {
-                    tracing::info!("regular web-login complete: {}", dto.account_name);
-                    let _ = watch_app.emit("regular-login-complete", dto);
-                }
-                Err(e) => {
-                    tracing::error!("regular web-login harvest failed: {e}");
-                    let _ = watch_app.emit("regular-login-error", format!("登入失敗: {e}"));
-                }
-            }
-            if let Some(win) = watch_app.get_webview_window("web-login") {
-                let _ = win.destroy();
-            }
-            return;
-        }
-        tracing::warn!("web-login: no login detected within timeout");
-    });
-
-    tracing::info!("regular web-login window opened");
-    Ok(())
+    webview_login::open_regular_web_login_window(app, session_id, account, password).await
 }
 
 /// Called by the web-login init script once login completes: harvest cookies,
@@ -1834,10 +845,10 @@ pub async fn regular_web_login_done(
     app: tauri::AppHandle,
     state: State<'_, AppState>,
 ) -> Result<(), ErrorDto> {
-    use tauri::Emitter;
+    use tauri::{Emitter, Manager};
 
     let ss = state.require_session(&session_id).await?;
-    match finalize_webview_login(
+    match webview_login::finalize_webview_login(
         &app,
         &ss,
         &session_id,
@@ -1864,436 +875,7 @@ pub async fn regular_web_login_done(
     Ok(())
 }
 
-/// Shared finalization for a webview-completed beanfun login: harvest ALL
-/// WebView2 cookies from `window_label`, inject them into the session's reqwest
-/// jar, build the [`Session`] with `account_name`, fetch game accounts, and
-/// store both. Returns the [`SessionDto`] or an error message.
-async fn finalize_webview_login(
-    app: &tauri::AppHandle,
-    ss: &std::sync::Arc<crate::models::session_state::SessionState>,
-    session_id: &str,
-    window_label: &str,
-    account_name: &str,
-    web_token_fallback: &str,
-    account_html: &str,
-) -> Result<SessionDto, String> {
-    let all_cookies = extract_webview2_cookies(app, window_label).await;
-    tracing::info!(
-        "web-login: extracted {} cookies from WebView2",
-        all_cookies.len()
-    );
-
-    let real_web_token = all_cookies
-        .iter()
-        .find(|(name, _, _, _)| name == "bfWebToken")
-        .map(|(_, value, _, _)| value.clone())
-        .unwrap_or_else(|| {
-            if web_token_fallback != "cookie_auth" && !web_token_fallback.is_empty() {
-                web_token_fallback.to_string()
-            } else {
-                String::new()
-            }
-        });
-
-    if real_web_token.is_empty() {
-        return Err("no bfWebToken found in webview cookies".to_string());
-    }
-
-    let tw_url: url::Url = "https://tw.beanfun.com/".parse().unwrap();
-    let login_url: url::Url = "https://login.beanfun.com/".parse().unwrap();
-    let newlogin_url: url::Url = "https://tw.newlogin.beanfun.com/".parse().unwrap();
-    for (name, value, domain, path) in &all_cookies {
-        let clean_domain = domain.trim_start_matches('.');
-        let jar_url =
-            if clean_domain.contains("login.beanfun.com") && !clean_domain.contains("newlogin") {
-                &login_url
-            } else if clean_domain.contains("newlogin") {
-                &newlogin_url
-            } else {
-                &tw_url
-            };
-        let path_str = if path.is_empty() { "/" } else { path.as_str() };
-        let cookie_str = format!("{}={}; Domain={}; Path={}", name, value, domain, path_str);
-        ss.cookie_jar.add_cookie_str(&cookie_str, jar_url);
-    }
-
-    let session = crate::models::session::Session {
-        token: real_web_token,
-        refresh_token: None,
-        expires_at: chrono::Utc::now() + chrono::Duration::hours(6),
-        region: crate::models::session::Region::TW,
-        account_name: account_name.to_string(),
-        session_key: None,
-        totp_state: None,
-    };
-
-    let accounts = crate::services::beanfun_service::get_game_accounts(
-        &ss.http_client,
-        &session,
-        &ss.cookie_jar,
-    )
-    .await
-    .unwrap_or_else(|e| {
-        tracing::warn!("web-login: reqwest get_game_accounts failed: {e}, using webview HTML");
-        crate::services::beanfun_service::parse_tw_account_list_html(account_html)
-    });
-    tracing::info!("web-login: got {} accounts", accounts.len());
-
-    let dto = SessionDto::from_session(&session, session_id);
-    *ss.session.write().await = Some(session);
-    *ss.game_accounts.write().await = accounts;
-    Ok(dto)
-}
-
-/// A cookie tuple: (name, value, domain, path).
-type CookieTuple = (String, String, String, String);
-
-/// Extract all cookies from a WebView2 window using the native CookieManager API.
-/// This reads HttpOnly cookies too (including secure/httponly flags).
-#[cfg(target_os = "windows")]
-async fn extract_webview2_cookies(app: &tauri::AppHandle, label: &str) -> Vec<CookieTuple> {
-    use std::sync::{Arc, Mutex};
-    use tauri::Manager;
-
-    let Some(webview) = app.get_webview_window(label) else {
-        tracing::warn!("GamePass: webview '{}' not found", label);
-        return Vec::new();
-    };
-
-    let cookies: Arc<Mutex<Vec<CookieTuple>>> = Arc::new(Mutex::new(Vec::new()));
-    let (tx, rx) = tokio::sync::oneshot::channel::<()>();
-    let tx = Arc::new(Mutex::new(Some(tx)));
-
-    let cookies_clone = cookies.clone();
-    let tx_clone = tx.clone();
-
-    let urls = vec![
-        "https://tw.beanfun.com".to_string(),
-        "https://login.beanfun.com".to_string(),
-        "https://tw.newlogin.beanfun.com".to_string(),
-    ];
-    let total = urls.len();
-
-    let result = webview.with_webview(move |wv| {
-        use webview2_com::{GetCookiesCompletedHandler, Microsoft::Web::WebView2::Win32::*};
-        use windows_core::Interface;
-
-        unsafe {
-            let controller = wv.controller();
-            let core: ICoreWebView2 = controller
-                .CoreWebView2()
-                .expect("failed to get CoreWebView2");
-
-            let core2: ICoreWebView2_2 = core.cast().expect("failed to cast to ICoreWebView2_2");
-            let cookie_manager: ICoreWebView2CookieManager =
-                core2.CookieManager().expect("failed to get CookieManager");
-
-            let remaining = Arc::new(Mutex::new(total));
-
-            for url_str in urls {
-                let cookies_ref = cookies_clone.clone();
-                let remaining_ref = remaining.clone();
-                let tx_ref = tx_clone.clone();
-
-                let handler =
-                    GetCookiesCompletedHandler::create(Box::new(move |hr, cookie_list| {
-                        if hr.is_ok() {
-                            if let Some(list) = cookie_list {
-                                let mut count: u32 = 0;
-                                let _ = list.Count(&mut count);
-                                for i in 0..count {
-                                    if let Ok(cookie) = list.GetValueAtIndex(i) {
-                                        let mut name = windows_core::PWSTR::null();
-                                        let mut value = windows_core::PWSTR::null();
-                                        let mut domain = windows_core::PWSTR::null();
-                                        let mut path = windows_core::PWSTR::null();
-
-                                        let _ = cookie.Name(&mut name);
-                                        let _ = cookie.Value(&mut value);
-                                        let _ = cookie.Domain(&mut domain);
-                                        let _ = cookie.Path(&mut path);
-
-                                        let n = name.to_string().unwrap_or_default();
-                                        let v = value.to_string().unwrap_or_default();
-                                        let d = domain.to_string().unwrap_or_default();
-                                        let p = path.to_string().unwrap_or_default();
-
-                                        // Free the PWSTR strings
-                                        if !name.is_null() {
-                                            windows_core::imp::CoTaskMemFree(name.as_ptr() as _);
-                                        }
-                                        if !value.is_null() {
-                                            windows_core::imp::CoTaskMemFree(value.as_ptr() as _);
-                                        }
-                                        if !domain.is_null() {
-                                            windows_core::imp::CoTaskMemFree(domain.as_ptr() as _);
-                                        }
-                                        if !path.is_null() {
-                                            windows_core::imp::CoTaskMemFree(path.as_ptr() as _);
-                                        }
-
-                                        if !n.is_empty() {
-                                            if let Ok(mut c) = cookies_ref.lock() {
-                                                c.push((n, v, d, p));
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-
-                        let mut rem = remaining_ref.lock().unwrap();
-                        *rem -= 1;
-                        if *rem == 0 {
-                            if let Some(sender) = tx_ref.lock().unwrap().take() {
-                                let _ = sender.send(());
-                            }
-                        }
-                        Ok(())
-                    }));
-
-                let url_hstring = windows_core::HSTRING::from(&url_str);
-                let _ = cookie_manager.GetCookies(&url_hstring, &handler);
-            }
-        }
-    });
-
-    if result.is_err() {
-        tracing::warn!("GamePass: with_webview failed");
-        return Vec::new();
-    }
-
-    // Wait for all cookie callbacks (timeout 10s)
-    let _ = tokio::time::timeout(std::time::Duration::from_secs(10), rx).await;
-
-    let result = cookies.lock().unwrap().clone();
-    tracing::info!("GamePass: extracted {} total cookies", result.len());
-    result
-}
-
-#[cfg(not(target_os = "windows"))]
-async fn extract_webview2_cookies(_app: &tauri::AppHandle, _label: &str) -> Vec<CookieTuple> {
-    Vec::new()
-}
-
-// ---------------------------------------------------------------------------
-// TW Regular (帳密) login — external reCAPTCHA helper window
-// ---------------------------------------------------------------------------
-//
-// The TW "Regular" (username + password) login on `login.beanfun.com` is
-// guarded by Google reCAPTCHA. Inside a bare WebView2 the widget frequently
-// fails to render — usually because `www.google.com` is unreachable for the
-// user. This helper opens the *official* login page in a small, frameless
-// window so the human can solve their own challenge; the injected script only
-// (a) swaps the reCAPTCHA origin to the Google-supported `recaptcha.net`
-// mirror so the widget loads, (b) trims the surrounding page chrome, and
-// (c) hands the token the human produced back to the backend. The challenge
-// is solved by a person — nothing here auto-solves or bypasses it.
-
-/// Window label for the external reCAPTCHA helper window.
-const RECAPTCHA_WINDOW_LABEL: &str = "recaptcha_window";
-
-/// Set right before the backend destroys the helper window after a SUCCESSFUL
-/// token capture, so the window-destroyed handler can tell that apart from a
-/// user closing the window. Without this, the destroy after phase 1 emits a
-/// spurious `recaptcha-cancelled` that can abort phase 2.
-static RECAPTCHA_DELIVERED: AtomicBool = AtomicBool::new(false);
-
-/// Set once a GamePass flow has been finalized, so the backend cookie poll and
-/// the JS `gamepass_webview_done` IPC (whichever wins the race) only complete the
-/// login once.
-static GAMEPASS_DONE: AtomicBool = AtomicBool::new(false);
-
-/// Consume the "token delivered" flag. `true` → the window closed because we
-/// captured a token (suppress cancel); `false` → user/close abort (emit cancel).
-pub fn recaptcha_take_delivered() -> bool {
-    RECAPTCHA_DELIVERED.swap(false, Ordering::SeqCst)
-}
-
-/// Injected at document-start into the reCAPTCHA helper window.
-///
-/// No `format!` placeholders here on purpose — it's a plain raw string so the
-/// JS braces/backticks need no escaping.
-const RECAPTCHA_INIT_SCRIPT: &str = r#"
-(() => {
-  'use strict';
-  // The reCAPTCHA token is origin-locked to login.beanfun.com, so the widget is
-  // solved on beanfun's own Login/Index page. Do NOT swap www.google.com for
-  // recaptcha.net: the Enterprise endpoint 404s on the mirror, leaving
-  // grecaptcha.render undefined. (Mainland support would need that wired up and
-  // verified separately.)
-
-  // Parity with the rest of the app's webviews.
-  try { Object.defineProperty(navigator, 'webdriver', { get: () => false }); } catch (e) {}
-
-  const OVERLAY_Z = 999999; // MUST stay below reCAPTCHA's challenge bframe (~2e9)
-
-  const onReady = () => {
-    // The helper window first passes through redirect pages (default.aspx →
-    // checkin → Login/Index). Only act on the actual login form.
-    if (location.href.indexOf('Login/Index') === -1) return;
-    console.log('[reCAPTCHA] on Login/Index, watching for token');
-
-    if (document.getElementById('__ov')) return;
-
-    // (1) Lay an opaque overlay + spinner from first paint so the user never
-    //     sees beanfun's raw login page flash by.
-    const style = document.createElement('style');
-    style.textContent =
-      '#__ov{position:fixed;inset:0;z-index:' + OVERLAY_Z + ';background:#1c1712;' +
-      'display:flex;flex-direction:column;align-items:center;justify-content:center;gap:18px;' +
-      'color:#f4ede4;font-family:system-ui,sans-serif;font-size:14px}' +
-      '#__sp{width:34px;height:34px;border-radius:50%;border:3px solid rgba(244,237,228,.25);' +
-      'border-top-color:#ff8201;animation:__r .8s linear infinite}' +
-      '@keyframes __r{to{transform:rotate(360deg)}}' +
-      '.grecaptcha-badge{z-index:' + (OVERLAY_Z + 1) + ' !important}';
-    const ov = document.createElement('div'); ov.id = '__ov';
-    const sp = document.createElement('div'); sp.id = '__sp';
-    const lb = document.createElement('div'); lb.textContent = '驗證載入中，請稍候…';
-    ov.appendChild(sp); ov.appendChild(lb);
-    (document.head || document.documentElement).appendChild(style);
-    (document.body || document.documentElement).appendChild(ov);
-
-    // (2) Detect the ANCHOR iframe (the "I'm not a robot" checkbox, not the
-    //     nine-grid bframe challenge).
-    const anchor = () =>
-      document.querySelector("iframe[src*='recaptcha'][src*='anchor']") ||
-      document.querySelector("iframe[title='reCAPTCHA']") ||
-      document.querySelector("iframe[src*='recaptcha']");
-
-    const findWidget = () => {
-      const a = anchor(); if (!a) return null;
-      let w = a.closest('.g-recaptcha');
-      if (!w) {
-        w = a;
-        for (let i = 0; i < 5 && w.parentElement && w.parentElement !== document.body; i++) {
-          w = w.parentElement;
-          if (w.offsetWidth >= 280 && w.offsetWidth <= 400) break;
-        }
-      }
-      return w;
-    };
-
-    let store = null;
-    try { store = window.sessionStorage; } catch (e) {}
-    const RKEY = '__rc_reloads__';
-
-    // (3) Once the widget exists, REPARENT it into the overlay. This is the only
-    //     reliable presentation: a pure z-index lift leaves the cross-origin
-    //     (out-of-process) iframe visible but unclickable (trapped in its
-    //     stacking context). Moving the iframe reloads it → the checkbox resets
-    //     to unchecked, but the user ticks the fresh one and getResponse() still
-    //     returns a fully valid token. If no widget ever renders, reveal the real
-    //     page (or, once the reload budget is spent, close so the frontend isn't
-    //     stuck on a blank helper — closing fires `recaptcha-cancelled`).
-    let placed = false;
-    // Reveal the real reCAPTCHA page (remove the opaque overlay). This is the
-    // graceful fallback for slow accelerator/VPN connections where the widget
-    // loads late or our selectors miss it: the user is NEVER left stuck on a
-    // spinner, and never depends on an IPC close that beanfun's CSP can block —
-    // they just solve the reCAPTCHA on the visible page (the token harvest below
-    // still works). Less pretty than the reparent, but always usable.
-    const reveal = () => {
-      if (placed) return;
-      placed = true; clearInterval(rt);
-      try { if (store) store.removeItem(RKEY); } catch (e) {}
-      ov.remove();
-    };
-    const place = () => {
-      if (placed) return;
-      const w = findWidget();
-      if (w && w !== document.body && w !== document.documentElement) {
-        placed = true; clearInterval(rt);
-        try { if (store) store.removeItem(RKEY); } catch (e) {}
-        w.style.position = 'relative';
-        w.style.zIndex = String(OVERLAY_Z + 2);
-        sp.remove();
-        lb.textContent = '請完成「我不是機器人」驗證';
-        ov.appendChild(w); // reparent → iframe reload → fresh checkbox
-      }
-    };
-    const rt = setInterval(() => { if (anchor()) place(); }, 200);
-
-    // (3b) Self-heal beanfun's first-load render race: ONLY if there is no
-    //      reCAPTCHA iframe at all after 4s (genuine render fail), reload once.
-    //      Slow connections keep the iframe present, so they skip this and just
-    //      wait for the widget → they don't get a disruptive reload.
-    setTimeout(() => {
-      if (!placed && !document.querySelector("iframe[src*='recaptcha']") && store && !store.getItem(RKEY)) {
-        store.setItem(RKEY, '1');
-        console.warn('[reCAPTCHA] no widget — reloading once');
-        location.reload();
-      }
-    }, 4000);
-
-    // Final fallback: if the widget still isn't reparented, reveal the real page
-    // rather than hide it behind the spinner forever.
-    setTimeout(reveal, 6500);
-
-    // (4) Harvest the Enterprise token once the human ticks the checkbox.
-    //     Primary source is grecaptcha.enterprise.getResponse() — beanfun is an
-    //     Enterprise checkbox and reads it straight into its XHR rather than into
-    //     a stable hidden input. Fall back to the response <textarea> / beanfun's
-    //     own field. We never fire beanfun's submit, so the token stays
-    //     unconsumed and can be replayed into AccountLogin's `Captcha` field.
-    const readToken = () => {
-      try {
-        const g = window.grecaptcha;
-        if (g && g.enterprise && typeof g.enterprise.getResponse === 'function') {
-          const t = g.enterprise.getResponse();
-          if (t) return t;
-        }
-        if (g && typeof g.getResponse === 'function') {
-          const t = g.getResponse();
-          if (t) return t;
-        }
-      } catch (e) {}
-      const el = document.getElementById('recaptcha-token') ||
-        document.querySelector('#g-recaptcha-response, textarea[name="g-recaptcha-response"]');
-      return el ? el.value : '';
-    };
-
-    const initial = readToken();
-    let done = false;
-    const timer = setInterval(() => {
-      if (done) return;
-      const val = readToken();
-      if (val && val !== initial && val.length > 50) {
-        done = true;
-        clearInterval(timer);
-        if (store) { try { store.removeItem(RKEY); } catch (e) {} }
-        const step = window.__RECAPTCHA_STEP__ || 'login';
-        // beanfun's page CSP blocks Tauri IPC (ipc.localhost → connect-src), so
-        // hand the token to the backend through a URL fragment it polls.
-        // Changing the hash isn't gated by connect-src and doesn't reload.
-        // reCAPTCHA tokens are URL-safe base64url, so '~' is a safe separator.
-        try { window.location.hash = 'mltoken=' + step + '~' + val; } catch (e) {}
-        // Best-effort IPC too, in case CSP allows it on some pages.
-        try {
-          if (window.__TAURI_INTERNALS__) {
-            window.__TAURI_INTERNALS__
-              .invoke('submit_login_token', { token: val, step })
-              .catch(() => {});
-          }
-        } catch (e) {}
-        console.log('[reCAPTCHA] token captured, handed to backend via fragment');
-      }
-    }, 500);
-  };
-
-  if (document.readyState === 'loading') {
-    document.addEventListener('DOMContentLoaded', onReady, { once: true });
-  } else {
-    onReady();
-  }
-})();
-"#;
-
 /// Open the external reCAPTCHA helper window for TW Regular (帳密) login.
-///
-/// Points a small, frameless, always-on-top WebView2 at the official login
-/// page and injects [`RECAPTCHA_INIT_SCRIPT`]. The human solves the challenge;
-/// the script invokes [`submit_login_token`] with the resulting token.
 ///
 /// `step` distinguishes the two challenge points (`"check"` for
 /// `CheckAccountType`, `"login"` for `AccountLogin`); it is echoed back in the
@@ -2304,225 +886,33 @@ pub async fn open_recaptcha_window(
     step: Option<String>,
     app: tauri::AppHandle,
 ) -> Result<(), ErrorDto> {
-    use tauri::WebviewWindowBuilder;
-
-    let step = step.unwrap_or_else(|| "login".to_string());
-
-    // Replace any existing helper window so we never stack two. Mark it as a
-    // backend close so its destroy doesn't emit a cancel that aborts this solve.
-    if let Some(existing) = app.get_webview_window(RECAPTCHA_WINDOW_LABEL) {
-        RECAPTCHA_DELIVERED.store(true, Ordering::SeqCst);
-        let _ = existing.destroy();
-        tokio::time::sleep(std::time::Duration::from_millis(200)).await;
-    }
-
-    // Tag the window with its step (consumed by RECAPTCHA_INIT_SCRIPT). The
-    // body's braces are safe here: it's interpolated as a value, not reparsed
-    // as a format string.
-    let init_script = format!("window.__RECAPTCHA_STEP__ = {step:?};\n{RECAPTCHA_INIT_SCRIPT}");
-
-    // Use a FRESH (incognito) profile every time. A persistent profile retains
-    // beanfun's login cookie, so loading Login/Index auto-completes the pSKey
-    // login and redirects to the portal — no reCAPTCHA is ever shown, no token
-    // is captured, and the frontend hangs on "登入中". A unique dir guarantees a
-    // clean login page with the reCAPTCHA every time.
-    let data_dir = std::env::temp_dir()
-        .join("MapleLink")
-        .join("recaptcha-incognito")
-        .join(uuid::Uuid::new_v4().to_string());
-    let _ = std::fs::create_dir_all(&data_dir);
-    let cleanup_dir = data_dir.clone();
-
-    // Navigate to bflogin/default.aspx — the server redirects to
-    // login.beanfun.com/Login/Index?pSKey={skey} (a bare Login/Index with no
-    // pSKey renders blank). Same entry point as open_gamepass_login.
-    let url = "https://tw.beanfun.com/beanfun_block/bflogin/default.aspx?service=999999_T0";
-
-    let window = WebviewWindowBuilder::new(
-        &app,
-        RECAPTCHA_WINDOW_LABEL,
-        tauri::WebviewUrl::External(url.parse().expect("static reCAPTCHA URL is valid")),
-    )
-    .title("Beanfun 驗證")
-    // Roomy enough that reCAPTCHA's image-challenge popup (~400px wide, opens
-    // beside the centred checkbox) isn't clipped on the right/bottom. Resizable
-    // so the user can enlarge further if a taller challenge appears.
-    .inner_size(500.0, 680.0)
-    .min_inner_size(440.0, 560.0)
-    .resizable(true)
-    // Keep the native title bar so the user can always close the window — a
-    // frameless window that fails to render leaves no way out and hangs the
-    // login. Closing fires `recaptcha-cancelled`, which unblocks the frontend.
-    .decorations(true)
-    .always_on_top(true)
-    .center()
-    .visible(true)
-    .user_agent(WEBVIEW_USER_AGENT)
-    .additional_browser_args("--disable-blink-features=AutomationControlled --no-sandbox")
-    .data_directory(data_dir)
-    .initialization_script(&init_script)
-    .build()
-    .map_err(|e| ErrorDto {
-        code: "AUTH_RECAPTCHA_WINDOW_FAILED".to_string(),
-        message: format!("Failed to open reCAPTCHA window: {e}"),
-        category: crate::models::error::ErrorCategory::Process,
-        details: None,
-    })?;
-
-    // WebView2 Tracking Prevention blocks google.com/gstatic third-party
-    // storage, which breaks reCAPTCHA's interaction (widget renders but clicks
-    // don't verify). Turn it off for this window's profile.
-    disable_tracking_prevention(&window);
-
-    // beanfun's page CSP blocks Tauri IPC (ipc.localhost), so the injected
-    // script hands the token back via the URL fragment (#mltoken=<step>~<token>).
-    // Poll the window URL for it.
-    let poll_app = app.clone();
-    tauri::async_runtime::spawn(async move {
-        // ~3 minutes at 500ms; the window-gone check ends it early on close.
-        for _ in 0..360 {
-            tokio::time::sleep(std::time::Duration::from_millis(500)).await;
-            let Some(win) = poll_app.get_webview_window(RECAPTCHA_WINDOW_LABEL) else {
-                return; // window closed or token already delivered
-            };
-            let Ok(u) = win.url() else { continue };
-            let Some(frag) = u.fragment() else { continue };
-            let Some(rest) = frag.strip_prefix("mltoken=") else {
-                continue;
-            };
-            if let Some((step, token)) = rest.split_once('~') {
-                if token.len() > 50 {
-                    deliver_recaptcha_token(&poll_app, token.to_string(), step.to_string());
-                    return;
-                }
-            }
-        }
-    });
-
-    // Remove the incognito profile once the window closes (best effort).
-    let cleanup_app = app.clone();
-    tauri::async_runtime::spawn(async move {
-        loop {
-            tokio::time::sleep(std::time::Duration::from_secs(1)).await;
-            if cleanup_app
-                .get_webview_window(RECAPTCHA_WINDOW_LABEL)
-                .is_none()
-            {
-                // Let WebView2 release its file locks before deleting.
-                tokio::time::sleep(std::time::Duration::from_secs(2)).await;
-                let _ = std::fs::remove_dir_all(&cleanup_dir);
-                break;
-            }
-        }
-    });
-
-    tracing::info!("reCAPTCHA helper window opened");
-    Ok(())
+    recaptcha_window::open_recaptcha_helper_window(app, step.unwrap_or_else(|| "login".to_string()))
+        .await
 }
-
-/// Disable WebView2 Tracking Prevention for a window's profile.
-///
-/// reCAPTCHA needs third-party storage/cookies on `google.com` / `gstatic.com`;
-/// Edge's Tracking Prevention blocks those by default, leaving the widget
-/// visible but non-functional. This is a profile-level setting (not a Chromium
-/// switch), so it must go through the WebView2 COM API.
-#[cfg(target_os = "windows")]
-fn disable_tracking_prevention(window: &tauri::WebviewWindow) {
-    use webview2_com::Microsoft::Web::WebView2::Win32::{
-        ICoreWebView2Profile3, ICoreWebView2_13, COREWEBVIEW2_TRACKING_PREVENTION_LEVEL_NONE,
-    };
-    use windows_core::Interface;
-
-    let result = window.with_webview(|wv| unsafe {
-        let Ok(core) = wv.controller().CoreWebView2() else {
-            return;
-        };
-        let Ok(core13) = core.cast::<ICoreWebView2_13>() else {
-            tracing::warn!(
-                "reCAPTCHA: ICoreWebView2_13 unavailable; cannot disable tracking prevention"
-            );
-            return;
-        };
-        let Ok(profile) = core13.Profile() else {
-            return;
-        };
-        let Ok(profile3) = profile.cast::<ICoreWebView2Profile3>() else {
-            tracing::warn!(
-                "reCAPTCHA: ICoreWebView2Profile3 unavailable; cannot disable tracking prevention"
-            );
-            return;
-        };
-        // Set on the profile (persisted in data_directory). The window goes
-        // through a redirect chain (default.aspx → checkin → Login/Index)
-        // before the reCAPTCHA loads, so this takes effect well before the
-        // widget initializes — no risky Reload needed.
-        match profile3
-            .SetPreferredTrackingPreventionLevel(COREWEBVIEW2_TRACKING_PREVENTION_LEVEL_NONE)
-        {
-            Ok(()) => tracing::info!("reCAPTCHA: tracking prevention disabled for helper window"),
-            Err(e) => tracing::warn!("reCAPTCHA: failed to disable tracking prevention: {e}"),
-        }
-    });
-
-    if result.is_err() {
-        tracing::warn!("reCAPTCHA: with_webview failed; tracking prevention left at default");
-    }
-}
-
-#[cfg(not(target_os = "windows"))]
-fn disable_tracking_prevention(_window: &tauri::WebviewWindow) {}
 
 /// Close the reCAPTCHA helper window (called by its own script when the widget
 /// fails to render, or as a frontend cleanup). Destroying it fires
 /// `recaptcha-cancelled`, which unblocks any pending login.
 #[tauri::command]
 pub async fn close_recaptcha_window(app: tauri::AppHandle) -> Result<(), ErrorDto> {
-    if let Some(win) = app.get_webview_window(RECAPTCHA_WINDOW_LABEL) {
-        let _ = win.destroy();
-    }
+    recaptcha_window::close_recaptcha_helper_window(&app);
     Ok(())
-}
-
-/// Payload re-emitted to the frontend when a reCAPTCHA token is captured.
-#[derive(Debug, Clone, serde::Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RecaptchaTokenEvent {
-    pub token: String,
-    /// `"check"` or `"login"` — which login step this token is for.
-    pub step: String,
-}
-
-/// Emit the captured reCAPTCHA token to the frontend and close the helper
-/// window. Shared by the fragment-poll path (primary) and the
-/// [`submit_login_token`] IPC command (fallback). Guards against double
-/// delivery by keying off the helper window still existing.
-fn deliver_recaptcha_token(app: &tauri::AppHandle, token: String, step: String) {
-    use tauri::Emitter;
-
-    // If the helper window is already gone, the token was already delivered.
-    let Some(win) = app.get_webview_window(RECAPTCHA_WINDOW_LABEL) else {
-        return;
-    };
-
-    tracing::info!(token_len = token.len(), step = %step, "reCAPTCHA token delivered");
-    if let Err(e) = app.emit("recaptcha-token", RecaptchaTokenEvent { token, step }) {
-        tracing::warn!("failed to emit recaptcha-token event: {e}");
-    }
-    // Mark this close as a successful delivery so on_window_event does NOT emit
-    // recaptcha-cancelled (which would abort the next phase).
-    RECAPTCHA_DELIVERED.store(true, Ordering::SeqCst);
-    let _ = win.destroy();
 }
 
 /// Fallback IPC path for receiving a reCAPTCHA token. Usually beanfun's page
 /// CSP blocks Tauri IPC, so the real delivery happens via the URL-fragment poll
-/// (see [`open_recaptcha_window`]); this remains for pages where IPC is allowed.
+/// (see [`recaptcha_window::open_recaptcha_helper_window`]); this remains for
+/// pages where IPC is allowed.
 #[tauri::command]
 pub async fn submit_login_token(
     token: String,
     step: Option<String>,
     app: tauri::AppHandle,
 ) -> Result<(), ErrorDto> {
-    deliver_recaptcha_token(&app, token, step.unwrap_or_else(|| "login".to_string()));
+    recaptcha_window::deliver_recaptcha_token(
+        &app,
+        token,
+        step.unwrap_or_else(|| "login".to_string()),
+    );
     Ok(())
 }
