+++ src-tauri/src/services/mod.rs
@@ -12,5 +12,9 @@ pub mod log_service;
 pub mod lr_service;
 pub mod network_service;
 pub mod process_service;
+pub mod recaptcha_window;
+pub mod session_key_fallback;
 pub mod update_service;
 pub mod web_launch;
+pub mod webview_login;
+pub mod webview_util;
